Support for align action.
Migrated from rive-cpp with these instructions:
https://2dimensions.slack.com/archives/CLLCU09T6/p1655746070344739

Adds support for the align action as discussed in these threads:
https://2dimensions.slack.com/archives/CHMAP278R/p1655506259768229
https://2dimensions.slack.com/archives/CHMAP278R/p1655592967543899

I'm specifically calling Actions anything that internally happens in response to a Listener (used to be just an input change but now will be generalized as the things listed in the diagram below) and an Event is a high-level external output from the StateMachine.

<img width="596" alt="listeners + internal actions" src="https://user-images.githubusercontent.com/454182/174460761-6ff41c51-1cb7-44ab-b265-d527931040f3.png">

Note that FireEvent isn't implemented yet but is called out as an example of another possible Action that a Listener will trigger.

Diffs=
9d1df4841 Cleanup.
aaad32b65 Support for align action.