Make LinearAnimationInstance copy-constructable
Earlier PR that added .reserve() was mean to fix a perceived fragile pointer issue, when we reallocated a std::vector containing (indirectly) a LinearAnimationInstance).

The actual bug was that we were relying on the default impl for copy-constructor, which was just failing to update our Counter. This PR makes that copy-constructor explicit (and removes the misleading comment above .reserve().

Note:

Now that we really understand the issue (just failing to update our counter in the copy-constructor) we *could* mark LinearAnimationInstance as non-copyable (as this PR does for StateMachineInstance) just for simplicity. With the .reserve() call, we don't *need* it to be copyable at the moment...

Diffs=
4f59a4362 Make LinearAnimationInstance copyable