Lua formatting
In one of the last viewer PRs I added in a lua formatter config file.

It works off of [LuaFormatter](https://github.com/Koihik/LuaFormatter) (think clang-format for lua).

It helps our premake5.lua files look more consistent and auto formats them for us.

The only trick to "help it along" is to scope filter, project, etc sections with lua do end blocks. This is purely cosmetic but I find it keeps the scope of the filter really clear. Similar to adding commas in dart to help the formatter reflow indentation.

Open to suggestions, I already I applied this style to the premake5_viewer.lua config file, but I included an example of the newly formatted premake5_tess.lua file with this PR.

Diffs=
412ab8f4d Lua formatting