coregraphics : strokes and gradients
Be sure to convert the stroke to a fillable path if we need to draw a stroke **with** a gradient.

CoreGraphics can natively draw a stroke (without having to explicitly set a clip), so we only do this convert-fill-to-stroke if we have to draw with a gradient.

This PR also fixes the case when the gradient needs to extend beyond its logical domain (i.e. clamp-tile). That fix is needed for fills as well as strokes.

Diffs=
c3f7325d8 Handle strokes and gradients