Force re-link viewer builds.
Our viewer has a few static lib dependencies like rive, rive_tess_renderer, etc. If you work in isolation on one of those dependencies, the viewer binary won't re-link. This was causing us to have to clean the build every time. The right solution is to make the build depend on any changes to those files. I couldn't find a good way to do that without merging all the premake files (we used to do this with dofile).

For now an easy solution is to just delete the viewer binary which forces a re-link any time you call build, which I think is ok.

Diffs=
c07c43a80 viewer premake depends on premake file for rive_tess
08c607bfe Force re-link viewer builds.