Add a simd library for SSE / NEON / WASM_SIMD
Adds a "simd" namespace with vector utilities (SSE, NEON, WASM_SIMD),
and implements RawPath::bounds in SIMD. This namespace relies
exclusively on the latest clang vector extensions, so from here on, Rive
needs to be built with a recent clang. On our own bots, we had to bump
the Android builder to NDK r25b, the Ubuntu builder to 22.04, and the
MacOS builder to macos-12.

RawPathBounds bench result:

  MacOS:   2.37ms -> .579 (4.1x)
  Windows: 3.53ms -> 1.68 (2.1x)

Windows SSE can be optimized down to .927ms (3.8x) by forcing the SSE
min/max instructions, but they have different behavior with NaN, which
is why clang doesn't use them directly, so it seems like an
over-optimization at this point.

Diffs=
986b49674 Add a simd library for SSE / NEON / WASM_SIMD (#4199)