Add dependency to correct grand parent for linear gradients.
The Text PR changed how we look-up a parent in the hierarchy to depend on for LinearGradient transforms. It introduced an issue where while we'd reference the right parent we wouldn't necessarily add the dependency to the correct one. We were using the current iterating node as the item to add the dependency to, which could be null if we didn't find one. In this case we still want to depend on the artboard.

Diffs=
686e5125b Add dependency to correct grand parent for linear gradients. (#4753)