fix up the viewer to progress time even when we miss a frame or two
basically sapp_frame_duration(); is not the amount of time since the last time frame(void) got called. and once we start missing the odd frame (easy to see on a 240hz monitor) we start losing track of time.

this is doing much better than without. i'm sure its not perfect, also my c++ is copy pasta and then a bit of 🔨 . in any case it doesnt look like a big deal at this point

Diffs=
6e95e7e1f fix up the viewer to progress time even when we miss a frame or two (#4912)