Rename reset to rewind.
In profiling Joel we saw that RawPath::reset was causing a lot of re-allocation. We realized the intention of reset in most of Rive code is to actually call rewind, so we renamed it and made other adjustments (like calling RawPath::rewind).

Diffs=
f4046e60b Rename reset to rewind. (#5004)