Fix clipping issue caused by wrong fillRule
Reported here: https://2dimensions.slack.com/archives/CSFM8K3CH/p1680123791322639

We were trying to be "smart", mostly for the benefit of the Tess renderer by sharing a Shape's Path when it was also used for clipping. With the Tess renderer it meant that a Path could be triangulated once and used for both clipping and rendering the actual path.

This breaks down when the Clip has a different fill rule from the referenced Shape, and is further complicated by the fact that the fill rules lives on the Shape's fills.

This PR simplifies things by just always giving clips their own render paths and not attempting to share them with drawable Shapes.

Diffs=
d9952ae77 Fix clipping issue caused by wrong fillRule (#5067)