Fix for runtime crash with skins nested in solo nodes.
Fixes a crash @HayesGordon caught when using the GameKit. Turns out it's a bug in the C++ runtime with loading of skins inside of nested solo nodes.

We used to resolve the parent of a skin in the "onAddedClean" cycle as it assumed onDirty would only be called after load. Solos can now call that during the "onAddedClean" cycle. We can do the same logic "onAddedDirty" and then the onDirty calls will only happen after we've ascertained we have a valid skinnable parent, so the assumptions still hold.

Diffs=
9266e0d30 Fix for runtime crash with skins nested in solo nodes. (#5109)