Introduce a DrawList to PLS
It's been great being able to draw everything in a single call, but we
are now at a point where we need separate draws for coarse
inner-path-fan tessellations and non-raster-ordered rendering. We will
probably also find separate draws superior to an uber shader for nested
clipping and raster images.

This change (finally) introduces a DrawList, which is capable of
recording all the types of draws we may need. Implementing the DrawList
also required incorporating EXT_base_instance into the GL backend, which
in turn necessitated a nice cleanup of our GL extension wrangling.

Diffs=
28c49dbbd Introduce a DrawList to PLS (#5432)

Co-authored-by: Chris Dalton <99840794+csmartdalton@users.noreply.github.com>