PLS cleanups
The PLS code has started to devolve into chaos. This PR is a start at attempting to de-spaghettify some of it.

It also fixes a bug that got introduced in bae069339. That PR assumed that interior triangulation did not write contour records to the GPU, but that's false. "outerCurve" patches still need a contour record in order to get the pathID.

Diffs=
60296f34f PLS cleanups (#5580)

Co-authored-by: Chris Dalton <99840794+csmartdalton@users.noreply.github.com>