Remove the 'EMIT_OFFSCREEN_VERTEX' macro in PLS
D3D and metal both have an opposite sense of Y when rendering to an
offscreen texture. This knowledge used to be baked into the shaders, but
that makes it difficult to correctly get triangle windings when we're
using face culling. This PR instead moves this knowledge of the Y
direction into variables by negating the Y viewport uniforms when
necessary.

This change also exposed an issue in minify.py where HLSL semantic names
were getting collisions because they are NOT case sensitive. So this
change also updates minify.py to not use lower case letters in
"@exported" names.

Diffs=
623766365 Remove the 'EMIT_OFFSCREEN_VERTEX' macro in PLS (#5615)

Co-authored-by: Chris Dalton <99840794+csmartdalton@users.noreply.github.com>