Rive Text for WASM Runtime
I had to bump some packages in order to work with the latest Node security updates. There were also some changes to how the pre-js code expects Modules to be named.

I also added some extra helpers to let you run the scripts directly and grab emsdk and premake if necessary.

The biggest and most important change is addition of Harfbuzz and SheenBidi to the wasm premake file. It also adds a new "clearGradient" call for Canvas 2D so that we can clear out a previously set gradient (we didn't need this previously as we never reset shaders to null).

Diffs=
ecbe56567 Rive Text for WASM Runtime (#5618)

Co-authored-by: Luigi Rosso <luigi-rosso@users.noreply.github.com>