A PLS flush should only invalidate the clip stack once
There was a bug where a flush would invalidate the clip stack indefinitely, meaning, after a flush, we would redraw the clip every single draw until it changed. This PR fixes it so we redraw the clip only once after a flush, and then go back to reusing it.

Diffs=
f601acae2 A PLS flush should only invalidate the clip stack once (#5844)

Co-authored-by: Chris Dalton <99840794+csmartdalton@users.noreply.github.com>