Fix follow path contention with MetricsPath
Building contour measures on the follow path constraint to remove contention on stored contour measures in the MetricsPath.

We were storing the contour data used for measuring on the MetricsPath. Calling computeLength with a different transform would mutate that stored contour data. So having the trim path and the FollowPathConstraint call computeLength with different transforms was causing problems as neither of them was guaranteed the contour data was for the correct transform they expected.

I fixed this by having the FollowPathConstraint store its own contours for the path.

Diffs=
033489f8b Fix follow path contention with MetricsPath (#5868)

Co-authored-by: Luigi Rosso <luigi-rosso@users.noreply.github.com>