Add a "clipRect" feature to PLS
The clipRect is a single rectangle whose coverage is determined analytically in the fragment shader, without the use of the clip buffer. Multiple rectangles can be intersected together to produce a single clipRect. Rectangles, including artboard clips, are detected automatically in PLSRenderer::clipPath().

This isn't always faster. On M1 it appears to be faster sometimes to just render the clipRect to the clip buffer instead of analyzing clipRect coverage at every pixel. But it does also allow us to use an artboard clip and a path clip together without nested clipping.

Diffs=
c6221a4a0 Add a "clipRect" feature to PLS (#5878)

Co-authored-by: Chris Dalton <99840794+csmartdalton@users.noreply.github.com>