Clean up some technical debt in PLSRenderer
There was a built-in assumption that the final path in an internal batch was not a clip update. Tidy up the code to remove this assumption, so that we can render clip paths for image meshes. While this change works to clean up the surrounding code, there is still plenty more leftover that we can tackle in future PRs.

Diffs=
f51298185 Clean up some technical debt in PLSRenderer (#5933)

Co-authored-by: Chris Dalton <99840794+csmartdalton@users.noreply.github.com>