Implement images and image meshes in WebGPU
This finalizes our WebGPU implementation so it renders our entire feature set and doesn't have random "not implemented" crashes.

Pad the image mesh uniform block to 256 bytes in order to satisfy WebGPU align restrictions.

Since we padded the image mesh uniform block, we can also modernize the GL backend to use uniform blocks.

Also update the spot where we sample the gradient texture to use 'textureLod()' instead of 'textureGrad()'. This more explicitly states what we want and may be faster.

This PR does *not* yet implement mipmaps in WebGPU, since there is not a built-in feature to do this. We will eventually need to roll out our own mipmap generator.

Diffs=
4f1a11b31 Implement images and image meshes in WebGPU (#6058)

Co-authored-by: Chris Dalton <99840794+csmartdalton@users.noreply.github.com>
Co-authored-by: Dragoș Tiselice <dragos@rive.app>