tendon crash fix
the current behavior of returning `StatusCode::FailedInversion;` when inversion fails was producing a crash. This was due to the fact that at the artboard level, we don't handle that error code in any specific way. So when a couple of lines later we call `onAddedDirty`, it tries to call `parent()->is<Skin>()`, and since parent() hasn't been set, it crashes.
This PR removes that error code and lets the app continue normally to mimic the editor behavior that allows for the inverse to fail, leaving it as an identity matrix.

Diffs=
a0f076e31 tendon crash fix (#6258)