PLS atomic mode
Implement an experimental PLS interlock mechanism that relies on atomics and barriers instead of raster ordering. This gets us working on the remaining Vulkan and D3D11+ hardware, enables direct rendering to the screen, and will probably end up being faster on a lot of platforms.

Because of how the atomic mode works, we can't fill paths with standard image textures anymore. On GL we introduce ARB_bindless_texture to work around this limitation when supported, but for other platforms, this PR also adds a new "imageRect" draw, which is a specialized, non-overlapping, antialiased implementation of drawImage() that is compatible with atomic mode.

Still TOOD:
- Simple reordering to limit barriers.
- Atomic mode on Metal for non-Apple Silicon.

Diffs=
35b2d8c0f PLS atomic mode (#6370)

Co-authored-by: Chris Dalton <99840794+csmartdalton@users.noreply.github.com>