make sure we force embedded assets when exporting for cloud renderer …
…preview

fixes an issue when hosted/referenced is set for assets for the cloud renderer.

by either producing a zip file or a file without the assets (both of which do not work well when previewing the render)

removed a few things that were flagged up as warnings as well.

i ignored an extra dependency for editor though:
```
  just_audio: ^0.9.36
```

gonna have a quick look if i can add a test...

Diffs=
acdaee21f make sure we force embedded assets when exporting for cloud renderer … (#6536)

Co-authored-by: Maxwell Talbot <talbot.maxwell@gmail.com>