propagate parent input change to nested input
We have decided to add a new attribute to our code generator that will allow to overwrite the behavior of property setters for the c++ runtime.
For the editor, the solution is more complex because we still need to reflect the nested artboard value in the inspector so users can follow changes both at the parent and the child level.

Diffs=
fee67f16d propagate parent input change to nested input (#6878)
ff2c0cbf6 apply final update if it has not been applied before (#6930)

Co-authored-by: hernan <hernan@rive.app>