treat cubic curve as quad when control point equals endpoint
using a follow path with oriented mode on, if the target path's last two points were a cubic bezier and a straight line, we were not calculating correctly the tangent at distance 100%.
That would cause the object with the follow path constraint to shift direction in the last position.
This PR fixes the issue by changing how that edge case works and calculates the tangent as a quadratic bezier curve instead.
fixes #6969

Diffs=
522a31bc2 treat cubic curve as quad when control point equals endpoint (#6919)

Co-authored-by: hernan <hernan@rive.app>