Handle NaN in PLS paths and transforms
Remove an assertion from Mat2D::mapBoundingBox() and instead make it explicitly return {0} when the points are empty or all NaN.

Introduce a "clipIsEmpty" boolean to the render stack so we can bail from draws early when the clip has NaN or empty paths. In the future we can take further advantage of this feature by marking the clip stack empty when its elements have an empty intersection.

Diffs=
7d0125c92 Handle NaN in PLS paths and transforms (#7176)

Co-authored-by: Chris Dalton <99840794+csmartdalton@users.noreply.github.com>