Xxxx support target align from position
A lot of files changed, but they're mostly boilerplate code and changes to the API of the actions.
This PR adds support for using Align Target without moving the aligned target to the pointer position.
Instead, it moves the element from its starting point following the mouse position.
It adds a new boolean property to enable the feature.
In order to achieve this, it's necessary to provide to all actions the current position and the previous position so the action can calculate the delta.

Diffs=
a55f1ffb6 Xxxx support target align from position (#7154)

Co-authored-by: hernan <hernan@rive.app>