Add required descriptor to avoid crash when creating a new RTV
CreateRenderTargetView doesn't allow a NULL descriptor when the target texture is typeless. This change adds a descriptor (with the assumption that the texture in question is a 2D texture in R8G8B8A8_UNORM format).

Authored-by: Elliot Colp <elliot.colp@gmail.com>

Diffs=
cd35e613d Add required descriptor to avoid crash when creating a new RTV (#7234)

Co-authored-by: Chris Dalton <99840794+csmartdalton@users.noreply.github.com>
Co-authored-by: Elliot Colp <elliot.colp@gmail.com>