Make ContourMeasure more robust
Make assertions and operations inside ContourMeasure more robust with degenerate data and NaN.

Add a rive::math::clamp() method with better behavior surrounding NaN.

Change the ContourMeasureIter constructor to accept a "const RawPath*". (The former definition with "const RawPath&" accepted r-values, which was unsafe since the iterator stored raw pointers directly to objects owned by the RawPath.)

Delete RawPath operator*(). The way this operator was defined, the matrix had to go on the wrong side of the path relative to how it operated on the data.

Diffs=
cd6210f42 Make ContourMeasure more robust (#7294)

Co-authored-by: Chris Dalton <99840794+csmartdalton@users.noreply.github.com>