Replace computeIntrinsicSize with measureLayout
Replaces computeIntrinsicSize with measureLayout which lets the layout engine call the measure function as necessary to allow the objects that can respond to changes in size to fit as best as they can given the constraints. This allow for better handling of line wrapping, ellipsis, and pushing content under text to the exact boundary of the text when using max width/height options in the layout engine.

I also tied in a few other fixes for how we allocate layout nodes and styles.

Diffs=
da0b71559 Replace computeIntrinsicSize with measureLayout (#7410)

Co-authored-by: Luigi Rosso <luigi-rosso@users.noreply.github.com>