use varuint for writing/reading objectid, total properties and property key in animation reset
objectId, propertykey are definitely written as varuint, not sure if properties total needs to be, but doesn't hurt to have it that way either I think.

I'm not sure if this is everything that needs to be updated. I tried to use `generate_core_runtime` to update the flutter runtime but that made a bunch of changes like unused imports all over the place...

Diffs=
07cbce551 use varuint for writing/reading objectid, total properties and property key in animation reset (#7510)

Co-authored-by: Arthur Vivian <arthur@rive.app>
Co-authored-by: Hernan Torrisi <hernantorrisi@gmail.com>