Xxxx improve hittest performance
In certain scenarios, there is no need to perform a hit test, so we precompute the conditions and early out from calculating the hit test.
If a shape has only listeners of type PointerDown and PointerUp, and is not an opaque target, it doesn't need to check for move events or exit events, which can save a lot of computations since it will skip most frames.

Diffs=
50bc398c4 Xxxx improve hittest performance (#7584)

Co-authored-by: hernan <hernan@rive.app>