add click event support
this PR adds click support to listeners.
Click is a new event type (like pointer down, pointer move).
It encompasses two stages (pointer down + pointer up) in the same object or group of objects that belong to the listener.
It processes all the phases of a click gesture
this PR also:
- guarantees that the click gesture is applied only once per frame (no double actions from overlapping shapes)
- supports starting the click gesture in one shape and ending it in another shape of the same listener (by promoting the hover state to the group and not on individual shapes)

Diffs=
405ca998b add click event support (#7668)

Co-authored-by: hernan <hernan@rive.app>