Use fixed function color blend in Vulkan
In atomic mode with no advanced blend, update Vulkan to use fixed
function color blend instead of an input attachment for color.

The biggest system-wide change is that we now have a
"fixedFunctionColorBlend" flag in pls::ShaderMiscFlags. This is the
correct way that we should have done this all along, and allows backends
to opt into the feature without requiring them to implicitly use it when
certian conditions are met.

The biggest Vulkan change here is that since this changes the entire
pipeline layout, we have to compile different fragment shaders. To do
this, we collapse the "spirv/*.vert" and "spirv/*.frag" variants of the
shaders into single ".main" files, then compile them up to 3 times --
once with `-DVERTEX`, once with `-DFRAGMENT`, and once with
`-DFIXED_FUNCTION_COLOR_BLEND -DFRAGMENT`.

Diffs=
295cf35fa Use fixed function color blend in Vulkan (#7753)

Co-authored-by: Chris Dalton <99840794+csmartdalton@users.noreply.github.com>