Fix static assertion in ResourceAllocationCounts
We can't assert the size of the struct is equal to the size of VecType because clang might allocate unused space in the vector. Insead ensure all the elements are of type size_t by asserting that the class is the size of kNumElements size_ts.

Diffs=
f8ff032a7 Fix static assertion in ResourceAllocationCounts (#7778)

Co-authored-by: Chris Dalton <99840794+csmartdalton@users.noreply.github.com>