added a blt command for render targets that do not support VK_IMAGE_USAGE_INPUT_ATTACHMENT_BIT
Currently, render targets that do not have the VK_IMAGE_USAGE_INPUT_ATTACHMENT_BIT flag will not work and just cause a crash. This adds an offscreen texture that can be used as an input attachment in its place and then a copy command to update the render target at the end.

Diffs=
8c7ac0329 added a blt command for render targets that do not support VK_IMAGE_USAGE_INPUT_ATTACHMENT_BIT (#7757)

Co-authored-by: Jonathon Copeland <jcopela4@gmail.com>