Add a "player" app to the tools
"player" is a tool that draws multiple copies of a .riv fullscreen and reports the FPS. It's (sort of) a simplification and rewrite of path_fiddle. "player" takes advantage of the built-in TestingWindow and TestHarness server so it can be easily deployed remotely. Currently, only android is supported as a remote target.

This PR also overhauls and simplifies the android_tools app to use NativeActivities, and deletes almost all the kotlin code.

https://github.com/user-attachments/assets/ff31cc6b-c7fb-42f0-8159-f7b1d9aad9b6

Diffs=
b172dca87 Add a "player" app to the tools (#8004)

Co-authored-by: Chris Dalton <99840794+csmartdalton@users.noreply.github.com>