pass wrap separately to support alignments
we need to pass the text wrap value separately in order to support right and center alignment.
Previously, since we were treating "no wrap" as an autosized text, the content wouldn't align to the text box and instead only align to itself.
This PR splits both as separate arguments and adds the new argument to the bindings.

Diffs=
451599d03 pass wrap separately to support alignments (#8058)
e32e2fde0 Start converting trickycubicstrokes to Rive (#8089)

Co-authored-by: Chris Dalton <99840794+csmartdalton@users.noreply.github.com>
Co-authored-by: blakdragan7 <jcopela4@gmail.com>
Co-authored-by: hernan <hernan@rive.app>