add bindable layout properties
Added many of the layout properties as bindable!
I included some properties that I think are currently not exposed to users, but it doesn't hurt to have them bindable already.
There are a couple more missing (like interpolation time) that I want to take a closer look. We should also decide whether we want them to be bindable at all.
The enum based properties are not bindable yet, we also need to discuss how that will work.

I  also had to run core generator again for the other files because there was a part missing, that's the reason for all the other core base file changes.

Diffs=
be2c46a3e add bindable layout properties (#8212)

Co-authored-by: hernan <hernan@rive.app>