Calculate LOD ahead of time for image paints
Since we don't do perspective transformations, the mipmap
level-of-detail for image paints is constant throughout the entire
shape. We can therefore save time by calculating the LOD ahead instead
of relying on shader derivatives in the fragment shader. This also saves
us from having to compute these derivatives explicitly outside of
control flow, which further improves performance.

To make room for LOD in the paintAuxBuffer, we also drop support for GL
bindless textures, which do not work similarly to descriptor indexing in
Vulkan, which we will use in the future if we decide this is a feature
we need.

Diffs=
7bef90845 Calculate LOD ahead of time for image paints (#8260)

Co-authored-by: Chris Dalton <99840794+csmartdalton@users.noreply.github.com>