enable all viewmodels as source for a condition
this extends state machine conditions to support any ViewModel property as the source of a condition instead of restricting it to the view model that is explicitly provided to the artboard.
This mimics the same behavior that artboard objects have, where they can target any ViewModel as the source for data binding.
This was surfaced from trying to build the fighters menu, where some properties are provided by the parent's data context, and not the component's data context itself.

Diffs=
5d5d418b7 enable all viewmodels as source for a condition (#8284)

Co-authored-by: hernan <hernan@rive.app>