Implement isHidden in DrawableProxy
Fixes an issue where nested artboards with hidden layouts that are also clipped would not render, and ultimately crash. In these cases, the draw calls in LayoutComponent were not running when isHidden() was true, but the layout's drawable proxy did not correctly implement isHidden, so it would try to render (incorrectly).

Diffs=
708a913ea Implement isHidden in DrawableProxy (#8302)

Co-authored-by: Philip Chung <philterdesign@gmail.com>