Optimize atomic rendering for input attachments
Pack clip data into an RGBA8 attachment and turn on src-over blending for both color and clip.

Handle advanced blend modes by rearranging the math such that the correct color isn't reached until *AFTER* the hardware blend state is applied.

This allows us to preserve clip and color contents by just emitting a=0 instead of loading the current value. It also saves flops by offloading the blending work onto the ROP blending unit, and serves as a hint to the hardware that it doesn't need to read or write anything when a == 0.

Diffs=
1b5e50fce Optimize atomic rendering for input attachments (#8310)

Co-authored-by: Chris Dalton <99840794+csmartdalton@users.noreply.github.com>