Make layout position type and scale type keyable
In order to support a request from Duo (this is actually something probably most users will want), we make layout position type (relative or absolute) keyable. By making a child layout absolute, it can break out of its parent's flexbox style layout system, which removes the "double gap" issue and also provides more freedom for the absolute layout to be repositioned, etc.

While doing this, it also became apparent that scale type (fixed, fill and hug) also needs to be keyable. This is because when switching from relative positioning with a fill scale type to absolute, we want to maintain sizing, so we would need to switch from auto (fill) to point (fixed) unit size. Again, this will provide additional flexibility for users to switch between the scale types in other scenarios.

Diffs=
638ee6434e Make layout position type and scale type keyable (#8316)

Co-authored-by: Philip Chung <philterdesign@gmail.com>