Upgrade legacy artboards to work with layouts
Files created before LayoutComponentStyle was added will contain Artboards that don't have a layout style applied. A layout style is required for the artboard to work with layouts. The resolution for this is to check if an Artboard has a layout component style at the time a LayoutNodeProvider (currently LayoutComponent or NestedArtboardLayout) is added to an Artboard. If a style doesn't exist, we create one.

Tested in the following ways:
1. Used a file from 3 years ago, verifying the Artboard has no LayoutComponentStyle, added layouts and nested artboard layouts, and verified that the Artboard now has a style, and its layout related inspectors behave as expected.
2. Created both layout components and nested artboard layouts on an Artboard with no style, verified it created the style, performed an undo, then verified the style was removed.

Diffs=
5cde901062 Upgrade legacy artboards to work with layouts (#8419)

Co-authored-by: Philip Chung <philterdesign@gmail.com>