editor: vector n-slicer should clamp all shapes contained
Previously, i thought it'd be nice to let vertices outside the bounds of the NSlicer to move relative to the bounds. You could get to this state by manually changing the intrinsic size to be smaller, so some vertices that used to be within bounds are now considered out.

However, I think this could be kinda confusing, not to mention harder to implement correctly (e.g. if the node is flipped, then the vertices outside the bounds also need to be flipped). I think not a lot of people will manually go and change the intrinsic size, so the simpler thing to do is to just clip everything:
<img width="1030" alt="image" src="https://github.com/user-attachments/assets/fc19caed-edce-4c15-bc3b-6d0e5f9496fc">

Diffs=
afcf0992da editor: vector n-slicer should clamp all shapes contained (#8428)

Co-authored-by: Susan Wang <susan@rive.app>