Add an experimental "clockwiseAtomic" rendering mode
Rather than relying on a resolve step, clockwiseAtomic mode finds coverage deltas at each fragment and incrementally blends them into the framebuffer.

Since the paint color of a single path is the same for all fragments at a single pixel, these incremental blends are commutative and the ordering of atomic deltas in the shader doesn't have to match the blend ordering.

In order to avoid barriers between paths, we allocate dedicated space in a global coverage buffer for each individual path to work out its atomics without collisions.

The end result is that, with a large enough coverage buffer, we are able to render an entire scene without barriers or resolves.

Diffs=
5a4418dba6 Add an experimental "clockwiseAtomic" rendering mode (#8442)

Co-authored-by: Chris Dalton <99840794+csmartdalton@users.noreply.github.com>