Use imageAsset size for layout
Fixes an issue only found in the web editor where images inside layouts do not render. A previous fix was to wait for the image asset to load and for the renderImage to become available, then update its scale using the layout size. However, on web, the renderImage's width/height remain 0/0 for a couple of frames after it is available.

The fix is to use the width/height from the imageAsset (core width/height) rather than the renderImage.

Diffs=
d60ea4cc71 Use imageAsset size for layout (#8462)

Co-authored-by: Philip Chung <philterdesign@gmail.com>