web: decode image on demand, not render
addresses issue discussed in these spots:

https://github.com/rive-app/rive/pull/8479
https://2dimensions.slack.com/archives/C07HQ4GS0BH/p1730744171360639

Does so by immediately decoding the image when requested by calling decodeImage on the factory (not waiting for first render to decode). As soon as the image is ready/decoded on the JS side we call in to cpp to tell it the image size and notify any asset referencers that the image has updated.

The decoded web image is stashed away and accessed on first actual render attempt so we can grab the correct webgl context to upload the texture with. This is the same as before but the only thing that is deferred is the actual final upload so size can be reported and used as soon as the image has been decoded.

https://github.com/user-attachments/assets/b7a4088f-3899-4b2b-bb03-e43cc4686012

Diffs=
446682cda9 web: decode image on demand, not render (#8484)

Co-authored-by: Luigi Rosso <luigi-rosso@users.noreply.github.com>