accept constraints without targets
we recently added validation in the runtime that discards objects if they're not valid.
And we were invalidating all constraints that didn't have targets
https://github.com/rive-app/rive/pull/8464/files

but some constraints work without targets attached to them.
This PR adds a new method that skips the validation if the constraint doesn't require a target to work.

Diffs=
f34822a124 accept constraints without targets (#8552)

Co-authored-by: hernan <hernan@rive.app>