feat: expose hit test to runtimes
This PR addresses a runtime limitation where we previously needed to trigger a pointer event to confirm if a listener existed at a specific location. With this update, users in game engines can choose whether pointer events in Rive should block raycasts from hitting items behind a Rive graphic. This change allows for hit testing without triggering the event itself.

I believe this is already available in Flutter, so this PR extends it to other runtimes as well

Diffs=
17474d3e2c feat: expose hit test to runtimes (#8598)

Co-authored-by: Adam <67035612+damzobridge@users.noreply.github.com>