remove letter spacing subtraction
texts in layouts with auto width are measured with no wrap applied. But when building the render style, the wrapping mode is respected.
This caused an inconsistency in the wrapping behavior because we were subtracting the letter spacing from the last character.
This removes that subtraction to keep it consistent.

Diffs=
1d2c9175d4 remove letter spacing subtraction (#8669)

Co-authored-by: hernan <hernan@rive.app>