do not use frameOrigin to validate if artboard has its origin translated
since we set the frameOrigin of nested artboards to false, we can't use that flag to validate whether the artboard's origin has been translated.
So we directly check for the originX and originY values to offset the pointer event.

Diffs=
d4a46062fb do not use frameOrigin to validate if artboard has its origin translated (#8788)

Co-authored-by: hernan <hernan@rive.app>