Ensure child layouts mark dirty when flex direction changes
There is some odd behavior when changing flexDirection values (row/column/etc). For example, when "Fill" scale type is being used, child layouts can become completely collapsed when the parent's flexDirection changes.

The heart of the issue was that since we compute things like flexGrow/Shrink and align values in syncStyle based on the parent's flexDirection (before passing them into the layout engine), when a parent layout's flexDirection changes, we need to mark it's direct layout children dirty so syncStyle will be reapplied on them (which includes LayoutComponents and NestedArtboardLayouts).

https://github.com/user-attachments/assets/2bb130a8-ecf2-46cb-a810-87fe135bbc0a

Diffs=
1be9b574d1 Ensure child layouts mark dirty when flex direction changes (#8792)

Co-authored-by: Philip Chung <philterdesign@gmail.com>