move advanced call after update
moving the triggers reset to the same place as inputs.
This is intended to fix an issue when triggers are called externally. And it mimics the current behavior.
In the future it might be good to explore a different approach to support reusing a trigger if the state machine doesn't settle in the first pass.

Diffs=
9d07b638c2 move advanced call after update (#8807)

Co-authored-by: hernan <hernan@rive.app>