only reset state change on new frame
Duo reported they are not getting values updated correctly.
Not sure if this is the (only) reason, but since we're resetting the state change on advance, if the state machine advances a second time in the same frame, it will reset its change status and could, down the line, not fire the state change notification.

This PR makes sure we only reset the state change if it's advancing on a new frame

Diffs=
88d24b0cd4 only reset state change on new frame (#8863)

Co-authored-by: hernan <hernan@rive.app>