initial work for testing data binding in the editor
This PR adds the infrastructure to make data binding testable in dart and the editor.
Given that .riv files lose a big part of their ids and some extra data, we need to remap elements and connect them back into ids to make them testable.
It makes use of the importers strategy that is used elsewhere.
It doesn't cover all cases, but it starts with triggers and state machine conditions.

Diffs=
5b9b388fe9 initial work for testing data binding in the editor (#8873)

Co-authored-by: hernan <hernan@rive.app>