fix sync with state machine triggers
this PR generalizes the fix that was done for input triggers to correctly be consumed only when the transition is selected. It fixes the same issue that was happening with view model triggers.
It also fixes scenarios where advanced was called before the state was consumed.
And another fix for dart and c++ syncing when advancing triggers.
Added some comments to the PR

Diffs=
641fe2ba15 fix sync with state machine triggers (#8950)

Co-authored-by: hernan <hernan@rive.app>