Improve feathered curve chopping
Now that feather joins are mathematically accurate, we can use them to approximate feathered curves. Regular stroke tessellations get subdivided by "polar segments" at fixed intervals in tangent angle; now we can similarly subdivide feathers using "polar joins" at fixed intervals in tangent angle. With this new strategy, feathers are finally perceptually identical to gaussian blurs.

NOTE: the work for this subdivision is still happening on the CPU. We will migrate this to the GPU once it has had some time to sit and prove that it's the final approach we want to commit to.

Diffs=
84f6941eb9 Improve feathered curve chopping (#8971)

Co-authored-by: Chris Dalton <99840794+csmartdalton@users.noreply.github.com>