change lerp
the previous lerp implementation could cause in certain scenarios that if t = 1, the output would be different than the input.
Because of floating point precision, `a - (b - a)` wasn't guaranteed to return `b`.
Down the line, when using trim paths for example, this small difference could cause the path not to draw properly.
Changing it to this implementation solves the issue.

Diffs=
1be3488d52 change lerp (#9000)

Co-authored-by: hernan <hernan@rive.app>