Fix Yoga crash with percent margin
When a layout is set to use percent margins at the same time that that layout node can't be computed due to a combination of hug/fill layouts, Yoga asserts with:

"availableWidth is indefinite so widthMeasureMode must be YGMeasureModeUndefined"
or
"availableHeight is indefinite so heightMeasureMode must be YGMeasureModeUndefined"

This crash only occurs with percent margins, not padding. At this point its unclear whether this is a bug in Yoga but this PR mitigates the crash.

Diffs=
339867b4de Fix Yoga crash with percent margin (#9010)

Co-authored-by: Philip Chung <philterdesign@gmail.com>