editor+runtime: allow setting tolerance to path measure
Before this PR, sometimes the movement on a path can be choppy, and this is noticeable everywhere I tested on mac, web, runtime:
https://github.com/user-attachments/assets/a396544a-40c1-4921-be11-01c978f50b4a
https://github.com/user-attachments/assets/78c54a00-b82f-42fe-ab85-f87961cccd62

This is because the default tolerance on the path measure is 0.5, which is too high. Making this 0.1 is smoother (tho probably less performant). I only changed this for text follow path, because it's more noticeable there.

After, on mac, web, runtime:
https://github.com/user-attachments/assets/99db7f81-27f6-46a9-b70c-bfe8ecb59933
https://github.com/user-attachments/assets/34b65a34-ff9e-40a0-a7c1-827146609d03
https://github.com/user-attachments/assets/be224730-dfba-4b1b-8d8a-c08dd48d8fa7

Diffs=
cfde2d5136 editor+runtime: allow setting tolerance to path measure (#9102)

Co-authored-by: Susan Wang <susan@rive.app>