editor+runtime: support text follow path with position
This PR can be reviewed by commits. It supports two main features:
1. I added a new boolean field, called 'radial', the difference is obvious when there are many lines in the text. False on the left, and True on the right:
![image](https://github.com/user-attachments/assets/16db8e55-f491-4c41-ba55-6e285577703b)

2. I also special cased 'position' when a text is following a path. Note this is a 'breaking' change, but from what I heard, folks unanimously like this new behavior. Before and after, for a x position modifier:
![image](https://github.com/user-attachments/assets/fbb131d1-7f78-4325-880f-b812085de902)

The editor UI for 'radial' is very simple for now. We should probably polish this up later once we've added all the features we want.
![image](https://github.com/user-attachments/assets/481eb7f8-dbe4-4206-8a09-5fc10bf40b9d)

I supported these features in both Dart and C++. I have a golden test on deck to commit, which looks like:
![image](https://github.com/user-attachments/assets/700e4dad-8d69-4432-b563-3af742f49e3e)

But I wanted to make sure stuff like the naming and approach are ok before I commit it, since there could be some churn.

Diffs=
39b813b46c editor+runtime: support text follow path with position (#9131)

Co-authored-by: Susan Wang <susan@rive.app>