editor+runtime: text follow path orient, strength, and trim path props
These properties take inspiration from Rive's other features. E.g. the creative time are used to manipulating the start+end+offset in trim path, and the follow path constraint has strength and orient.

For trim path, I'm noticing that the stroke always wraps. However, I thought it'd be kinda cool to intentionally not wrap if the range is <100% on a closed path. Then you can get a cool kinda effect with tangents (e.g. the penultimate row on the golden test below).

This is the new UI (in a popout):
![image](https://github.com/user-attachments/assets/952e0bc0-8c0e-47b0-8f6a-03ba67eb6349)

I added a golden test that looks like this:
<img width="926" alt="image" src="https://github.com/user-attachments/assets/376b7969-ea7b-479f-8bad-d098e5914bf0" />

Diffs=
7fe2e81969 editor+runtime: text follow path orient, strength, and trim path props (#9177)

Co-authored-by: Susan Wang <susan@rive.app>