n-slicing: update path when removing n-slicing
The Path needs to recompute its render path to update the positions of its vertices on the screen.

When an NSlicedNode is becoming irrelevant to a Path, e.g. when we ungroup it, or when we drag a shape out of the hierarchy, Path needs to update.

However, currently it doesn't because update() happens after a core component's onRemoved. This means, deformer from the Path's perspective will be null.

This PR tells Path to update whenever Dirt.nslicer is propagated, regardless of whether it's currently under the effect of one.
But it will only recalculate path on worldTransform dirt if a deformer is valid.

I verified that when we ungroup or drag a shape out of a NSlicedNode, the most updated path is shown.

Diffs=
cecd189934 n-slicing: update path when removing n-slicing  (#9211)

Co-authored-by: Susan Wang <susan@rive.app>