Overhaul Vulkan synchronization
The Vulkan backend has been requiring the client to pass fences along from queue submissions, and then it waits on these fences on its own timeline. This is hard to work with, bug prone, not optimal for performance, and just isn't very Vulkany.

This change removes all CPU-GPU synchronization from Rive's backend. Instead, the client is responsible for waiting on its own fences (which, realistically, it was already doing anyway). Instead of passing in fences, it just passes in two lifecycle counters currentFrameNumber and safeFrameNumber. Resources used during the flush belong to 'currentFrameNumber'. Resources last used on or before 'safeFrameNumber' are safe for Rive to be release or recycle.

Diffs=
335ed61d42 Overhaul Vulkan synchronization (#9280)

Co-authored-by: Chris Dalton <99840794+csmartdalton@users.noreply.github.com>