Fix Layout issue with image when N-slice applied
# Description
Images were not respecting their parent layout's size when an N-slice was was applied as a mesh (runtime only).

# Details
This bug was introduced as part of another fix (https://github.com/rive-app/rive/pull/8840) where the correct value of type was not being applied for n-slice meshes. Added a golden test

Diffs=
c118df3ccb Fix Layout issue with image when N-slice applied (#9283)

Co-authored-by: Philip Chung <philterdesign@gmail.com>