Fix for layoutstyle dirt continually being added in nested artboards
# Description
Noticed when using scrollbar constraints in a nested artboard that LayoutStyle can continue to be dirtied.

# Details
Continual LayoutStyle dirt was the result of a recent change related to Layout Direction. With layout direction, we needed the cascade of direction information to result in calculateLayout and updateLayoutBounds, so we forced that to happen in nested artboards. However, this caused a side effect with continual dirt. The fix now checks to see if the cascade resulted in any changes, in which case we call calculateLayout and updateLayoutBounds.

Diffs=
05d87692bb Fix for layoutstyle dirt continually being added in nested artboards (#9298)

Co-authored-by: Philip Chung <philterdesign@gmail.com>