add support for joystick time based dependents
# Description
fix for joystick controlled timelines not updating the time

# Details
If a state machine has a timeline controlling a nested remapped animation through keyframes,
and that timeline is itself controlled by a joystick
our update cycle would not apply the time updated by the joystick because the advance cycle has passed.
This PR looks for joystick controlled animations, and stores which objects are nested remapped animations to advance them while they are being applied.

Diffs=
038bd34aee add support for joystick time based dependents (#9272)

Co-authored-by: hernan <hernan@rive.app>