Fix rendering corruption from nested clipIDs
In an effort to reduce varying counts, we were stuffing the outerClipID into a float (v_paint.r), and then using this value to compare with a half (v_clipID). On ANGLE at least, this was leading to corruption because even if v_clipID == v_outerClipID in the vertex shader, they were no longer equal by the time they reached the fragment shader.

Adding one more varying and passing v_outerClipID through as a half instead of a float fixes this problem for now. In the future we may want to consider making them "flat short" to be even safer, but the performance cost will be relatively high on some platforms.

The current solution comes at a slight perf cost when not using nested clipping, but appears to conversely improve things when nested clipping is used.

marty (no nested clipping): 132.9 fps -> 132.5
popsicle_loader_bite_4 (heavy nested clipping): 131.9 fps -> 132.9

Diffs=
373238831d Fix rendering corruption from nested clipIDs (#9383)

Co-authored-by: Chris Dalton <99840794+csmartdalton@users.noreply.github.com>