refactor: Work out low level barriers at the RenderContext level
Determine all the barriers at a higher level so that the individual backends don't have to duplicate this work.

This also prepares us for another refactor that will move the blend state into PipelineState, and to start supporting "blend barriers" that we can use with non-coherent advanced blend extensions.

Diffs=
0ebca777f4 refactor: Work out low level barriers at the RenderContext level (#9474)

Co-authored-by: Chris Dalton <99840794+csmartdalton@users.noreply.github.com>