library: allow loading library that uses another library
This PR lazily decodes libraries referred to by other libraries.
Fixes: https://github.com/rive-app/rive/issues/9457

Before this PR, a host file will only load the FileAssets referred in itself. If that FileAsset is a LibraryAsset, we don't go look for other libraries it refers. This PR changes that to lazily loading whichever LibraryAssets are needed in the export process.

This is manually verified and I also wrote a test for it.

Diffs=
bcdfcda330 library: allow loading library that uses another library (#9480)

Co-authored-by: Susan Wang <susan@rive.app>