Update our renderer premake vulkan version to 1.4.321 (plus another build fix) (#10452) c26e82adc9 * Update our renderer premake vulkan version to 1.4.321 (and also add a few missing RIVE_UNREACHABLEs)
Was getting build errors trying to build Vulkan relating to the vk memory allocator header noting a mismatch between the vulkan sdk being reported and the #defines in the headers, and Chris mentioned that we could just upgrade it to be 1.4 now.

Additionally, I got a couple errors in two functions that didn't return in the "default" case, which needed RIVE_UNREACHABLE() at the end like the other similar functions had.

* Rebase the pixel6 GL and Vulkan goldens

Co-authored-by: Josh Jersild <joshua@rive.app>