fix(gl): Limit the tess/grad instance counts on Mali/PowerVR (#10520) b0646c5c97 fix(gl): Limit the tess/grad instance conts on Mali/PowerVR
Limit the number of instances we draw at a time when rendering the
tessellation and gradient textures. PowerVR and ARM GPUs have been known
to crash when rendering large instance counts. We're already limiting
the instance counts for main draws, but we missed the tessellation and
gradient passes when we added this workaround.

Co-authored-by: Chris Dalton <99840794+csmartdalton@users.noreply.github.com>