fix(vk): Properly preserve render targets when using MSAA (#10630) 5df1a42463 * fix(vk): Properly preserve render targets when using MSAA
Unfortunately, Vulkan does not provide a mechanism to initialize our
(transient) MSAA texture with the contents of the (non-MSAA)
renderTarget texture at the beginning of a render pass. To work around
this limitation while supporting LoadAction::preserveRenderTarget, we
literally draw the previous render target contents into the framebuffer
at the beginning of the pass.

* Fix one more barrier

Co-authored-by: Chris Dalton <99840794+csmartdalton@users.noreply.github.com>